+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [2.3.0] - 2022-03-06
+#### Feature
+- Support for match presets
+- Endpoint for adding tournament groups
+
+#### Changed
+- Reverse order of players on rematch
+
+#### Fixed
+- Don't allow numbers to be stolen for Tic-Tac-Toe
+
 ## [2.2.0] - 2021-12-04
 #### Feature
 - Smartcard `UID` support for each player
@@ -79,6 +90,7 @@
 #### Feature
 - Intial version of API for [kcapp-frontend](https://github.com/kcapp/frontend)
 
+[2.3.0]: https://github.com/kcapp/api/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/kcapp/api/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/kcapp/api/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/kcapp/api/compare/v1.2.0...v2.0.0